Repository Restructuring
Moved old partial files to their own "partial" folder to give the repository
more clarity.

Removed editor files (.ipynb_checkpoints/) and OS files (.DS_Store). Also
created a .gitignore to not commit them in the future.

Added the archives used in the data analysis back in 2023, mainly so if we need
to run it again or accidentally remove the files, the origial .html file is
still retrievable.