Writing and Code Format Improvements
Massive improvements to the quality of writing in the tutorial. Notably the
structure of the writing is better, and everything just flows better.

Code formatting has been improved by improving the spacing of the code. Actual
functioning of the code has been left mostly unchanged, save for a section to
save the box office data to a file, allowing for the paper to be reproduced.